Move calendarjs endpoints
