+++ docs/announcements.api.v2.rst
@@ -6,16 +6,16 @@ announcements.api.v2 package
    :undoc-members:
    :show-inheritance:
 
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   announcements.api.v2.serializers
+
 Submodules
 ----------
-
-announcements.api.v2.serializers module
----------------------------------------
-
-.. automodule:: announcements.api.v2.serializers
-   :members:
-   :undoc-members:
-   :show-inheritance:
 
 announcements.api.v2.urls module
 --------------------------------
