+++ docs/announcements.api.v2.serializers.rst
@@ -1,0 +1,26 @@
+announcements.api.v2.serializers package
+========================================
+
+.. automodule:: announcements.api.v2.serializers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+announcements.api.v2.serializers.frontpage\_article module
+----------------------------------------------------------
+
+.. automodule:: announcements.api.v2.serializers.frontpage_article
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+announcements.api.v2.serializers.slide module
+---------------------------------------------
+
+.. automodule:: announcements.api.v2.serializers.slide
+   :members:
+   :undoc-members:
+   :show-inheritance:
