+++ docs/members.api.calendarjs.rst
@@ -1,0 +1,26 @@
+members.api.calendarjs package
+==============================
+
+.. automodule:: members.api.calendarjs
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+members.api.calendarjs.serializers module
+-----------------------------------------
+
+.. automodule:: members.api.calendarjs.serializers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+members.api.calendarjs.views module
+-----------------------------------
+
+.. automodule:: members.api.calendarjs.views
+   :members:
+   :undoc-members:
+   :show-inheritance:
