+++ docs/members.api.v1.rst
@@ -1,0 +1,34 @@
+members.api.v1 package
+======================
+
+.. automodule:: members.api.v1
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+members.api.v1.serializers module
+---------------------------------
+
+.. automodule:: members.api.v1.serializers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+members.api.v1.urls module
+--------------------------
+
+.. automodule:: members.api.v1.urls
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+members.api.v1.viewsets module
+------------------------------
+
+.. automodule:: members.api.v1.viewsets
+   :members:
+   :undoc-members:
+   :show-inheritance:
