+++ docs/thaliawebsite.api.calendarjs.rst
@@ -1,0 +1,18 @@
+thaliawebsite.api.calendarjs package
+====================================
+
+.. automodule:: thaliawebsite.api.calendarjs
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+thaliawebsite.api.calendarjs.serializers module
+-----------------------------------------------
+
+.. automodule:: thaliawebsite.api.calendarjs.serializers
+   :members:
+   :undoc-members:
+   :show-inheritance:
