+++ website/announcements/api/v2/serializers/__init__.py
@@ -1,0 +1,2 @@
+from .frontpage_article import FrontpageArticleSerializer
+from .slide import SlideSerializer
