+++ website/announcements/api/v2/serializers/frontpage_article.py
@@ -1,0 +1,16 @@
+from rest_framework import serializers
+
+from announcements.models import FrontpageArticle
+from thaliawebsite.api.v2.serializers.thumbnail import CleanedHTMLSerializer
+
+
+class FrontpageArticleSerializer(serializers.ModelSerializer):
+    """FrontpageArticle serializer."""
+
+    class Meta:
+        """Meta class for the serializer."""
+
+        model = FrontpageArticle
+        fields = ("pk", "title", "content")
+
+    content = CleanedHTMLSerializer()
