--- website/events/api/serializers/base/__init__.py
@@ -1,1 +1,0 @@
-from .calendarjs import *
