+++ website/events/api/v1/viewsets/__init__.py
@@ -1,3 +1,2 @@
-from .base import *
 from .events import *
 from .event_registrations import *
