+++ website/members/api/v1/urls.py
@@ -1,7 +1,7 @@
 """DRF routes defined by the members package"""
 from rest_framework import routers
 
-from members.api import viewsets
+from . import viewsets
 
 router = routers.SimpleRouter()
 router.register(r"members", viewsets.MemberViewset)
